+++ model/context/modifier.go
@@ -6,12 +6,12 @@ import (
 )
 
 type ModifierContext struct {
-	addLocalDirectAttackModifiers kv.Map[uint, []modifier.Modifier[DamageContext]]
-	addLocalFinalAttackModifiers  kv.Map[uint, []modifier.Modifier[DamageContext]]
-	addLocalDefenceModifiers      kv.Map[uint, []modifier.Modifier[DamageContext]]
-	addLocalChargeModifiers       kv.Map[uint, []modifier.Modifier[ChargeContext]]
-	addLocalHealModifiers         kv.Map[uint, []modifier.Modifier[HealContext]]
-	addLocalCostModifiers         kv.Map[uint, []modifier.Modifier[CostContext]]
+	addLocalDirectAttackModifiers kv.Map[uint64, []modifier.Modifier[DamageContext]]
+	addLocalFinalAttackModifiers  kv.Map[uint64, []modifier.Modifier[DamageContext]]
+	addLocalDefenceModifiers      kv.Map[uint64, []modifier.Modifier[DamageContext]]
+	addLocalChargeModifiers       kv.Map[uint64, []modifier.Modifier[ChargeContext]]
+	addLocalHealModifiers         kv.Map[uint64, []modifier.Modifier[HealContext]]
+	addLocalCostModifiers         kv.Map[uint64, []modifier.Modifier[CostContext]]
 
 	addGlobalDirectAttackModifiers []modifier.Modifier[DamageContext]
 	addGlobalFinalAttackModifiers  []modifier.Modifier[DamageContext]
@@ -20,12 +20,12 @@ type ModifierContext struct {
 	addGlobalHealModifiers         []modifier.Modifier[HealContext]
 	addGlobalCostModifiers         []modifier.Modifier[CostContext]
 
-	removeLocalDirectAttackModifiers kv.Map[uint, []modifier.Modifier[DamageContext]]
-	removeLocalFinalAttackModifiers  kv.Map[uint, []modifier.Modifier[DamageContext]]
-	removeLocalDefenceModifiers      kv.Map[uint, []modifier.Modifier[DamageContext]]
-	removeLocalChargeModifiers       kv.Map[uint, []modifier.Modifier[ChargeContext]]
-	removeLocalHealModifiers         kv.Map[uint, []modifier.Modifier[HealContext]]
-	removeLocalCostModifiers         kv.Map[uint, []modifier.Modifier[CostContext]]
+	removeLocalDirectAttackModifiers kv.Map[uint64, []modifier.Modifier[DamageContext]]
+	removeLocalFinalAttackModifiers  kv.Map[uint64, []modifier.Modifier[DamageContext]]
+	removeLocalDefenceModifiers      kv.Map[uint64, []modifier.Modifier[DamageContext]]
+	removeLocalChargeModifiers       kv.Map[uint64, []modifier.Modifier[ChargeContext]]
+	removeLocalHealModifiers         kv.Map[uint64, []modifier.Modifier[HealContext]]
+	removeLocalCostModifiers         kv.Map[uint64, []modifier.Modifier[CostContext]]
 
 	removeGlobalDirectAttackModifiers []modifier.Modifier[DamageContext]
 	removeGlobalFinalAttackModifiers  []modifier.Modifier[DamageContext]
@@ -35,8 +35,8 @@ type ModifierContext struct {
 	removeGlobalCostModifiers         []modifier.Modifier[CostContext]
 }
 
-func (m *ModifierContext) AddLocalDirectAttackModifier(targetCharacter uint, handler modifier.Modifier[DamageContext]) {
-	var null kv.Map[uint, []modifier.Modifier[DamageContext]] = nil
+func (m *ModifierContext) AddLocalDirectAttackModifier(targetCharacter uint64, handler modifier.Modifier[DamageContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[DamageContext]] = nil
 	if m.addLocalDirectAttackModifiers != null {
 		modifiers := append(m.addLocalDirectAttackModifiers.Get(targetCharacter), handler)
 		m.addLocalDirectAttackModifiers.Set(targetCharacter, modifiers)
@@ -47,8 +47,8 @@ func (m *ModifierContext) AddLocalDirectAttackModifier(targetCharacter uint, han
 	}
 }
 
-func (m *ModifierContext) AddLocalFinalAttackModifier(targetCharacter uint, handler modifier.Modifier[DamageContext]) {
-	var null kv.Map[uint, []modifier.Modifier[DamageContext]] = nil
+func (m *ModifierContext) AddLocalFinalAttackModifier(targetCharacter uint64, handler modifier.Modifier[DamageContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[DamageContext]] = nil
 	if m.addLocalFinalAttackModifiers != null {
 		modifiers := append(m.addLocalFinalAttackModifiers.Get(targetCharacter), handler)
 		m.addLocalFinalAttackModifiers.Set(targetCharacter, modifiers)
@@ -59,8 +59,8 @@ func (m *ModifierContext) AddLocalFinalAttackModifier(targetCharacter uint, hand
 	}
 }
 
-func (m *ModifierContext) AddLocalDefenceModifier(targetCharacter uint, handler modifier.Modifier[DamageContext]) {
-	var null kv.Map[uint, []modifier.Modifier[DamageContext]] = nil
+func (m *ModifierContext) AddLocalDefenceModifier(targetCharacter uint64, handler modifier.Modifier[DamageContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[DamageContext]] = nil
 	if m.addLocalDefenceModifiers != null {
 		modifiers := append(m.addLocalDefenceModifiers.Get(targetCharacter), handler)
 		m.addLocalDefenceModifiers.Set(targetCharacter, modifiers)
@@ -71,8 +71,8 @@ func (m *ModifierContext) AddLocalDefenceModifier(targetCharacter uint, handler 
 	}
 }
 
-func (m *ModifierContext) AddLocalChargeModifier(targetCharacter uint, handler modifier.Modifier[ChargeContext]) {
-	var null kv.Map[uint, []modifier.Modifier[ChargeContext]] = nil
+func (m *ModifierContext) AddLocalChargeModifier(targetCharacter uint64, handler modifier.Modifier[ChargeContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[ChargeContext]] = nil
 	if m.addLocalChargeModifiers != null {
 		modifiers := append(m.addLocalChargeModifiers.Get(targetCharacter), handler)
 		m.addLocalChargeModifiers.Set(targetCharacter, modifiers)
@@ -83,8 +83,8 @@ func (m *ModifierContext) AddLocalChargeModifier(targetCharacter uint, handler m
 	}
 }
 
-func (m *ModifierContext) AddLocalHealModifier(targetCharacter uint, handler modifier.Modifier[HealContext]) {
-	var null kv.Map[uint, []modifier.Modifier[HealContext]] = nil
+func (m *ModifierContext) AddLocalHealModifier(targetCharacter uint64, handler modifier.Modifier[HealContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[HealContext]] = nil
 	if m.addLocalHealModifiers != null {
 		modifiers := append(m.addLocalHealModifiers.Get(targetCharacter), handler)
 		m.addLocalHealModifiers.Set(targetCharacter, modifiers)
@@ -95,8 +95,8 @@ func (m *ModifierContext) AddLocalHealModifier(targetCharacter uint, handler mod
 	}
 }
 
-func (m *ModifierContext) AddLocalCostModifier(targetCharacter uint, handler modifier.Modifier[CostContext]) {
-	var null kv.Map[uint, []modifier.Modifier[CostContext]] = nil
+func (m *ModifierContext) AddLocalCostModifier(targetCharacter uint64, handler modifier.Modifier[CostContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[CostContext]] = nil
 	if m.addLocalCostModifiers != null {
 		modifiers := append(m.addLocalCostModifiers.Get(targetCharacter), handler)
 		m.addLocalCostModifiers.Set(targetCharacter, modifiers)
@@ -155,8 +155,8 @@ func (m *ModifierContext) AddGlobalCostModifier(handler modifier.Modifier[CostCo
 	}
 }
 
-func (m *ModifierContext) RemoveLocalDirectAttackModifier(targetCharacter uint, handler modifier.Modifier[DamageContext]) {
-	var null kv.Map[uint, []modifier.Modifier[DamageContext]] = nil
+func (m *ModifierContext) RemoveLocalDirectAttackModifier(targetCharacter uint64, handler modifier.Modifier[DamageContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[DamageContext]] = nil
 	if m.removeLocalDirectAttackModifiers != null {
 		modifiers := append(m.removeLocalDirectAttackModifiers.Get(targetCharacter), handler)
 		m.removeLocalDirectAttackModifiers.Set(targetCharacter, modifiers)
@@ -167,8 +167,8 @@ func (m *ModifierContext) RemoveLocalDirectAttackModifier(targetCharacter uint, 
 	}
 }
 
-func (m *ModifierContext) RemoveLocalFinalAttackModifier(targetCharacter uint, handler modifier.Modifier[DamageContext]) {
-	var null kv.Map[uint, []modifier.Modifier[DamageContext]] = nil
+func (m *ModifierContext) RemoveLocalFinalAttackModifier(targetCharacter uint64, handler modifier.Modifier[DamageContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[DamageContext]] = nil
 	if m.removeLocalFinalAttackModifiers != null {
 		modifiers := append(m.removeLocalFinalAttackModifiers.Get(targetCharacter), handler)
 		m.removeLocalFinalAttackModifiers.Set(targetCharacter, modifiers)
@@ -179,8 +179,8 @@ func (m *ModifierContext) RemoveLocalFinalAttackModifier(targetCharacter uint, h
 	}
 }
 
-func (m *ModifierContext) RemoveLocalDefenceModifier(targetCharacter uint, handler modifier.Modifier[DamageContext]) {
-	var null kv.Map[uint, []modifier.Modifier[DamageContext]] = nil
+func (m *ModifierContext) RemoveLocalDefenceModifier(targetCharacter uint64, handler modifier.Modifier[DamageContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[DamageContext]] = nil
 	if m.removeLocalDefenceModifiers != null {
 		modifiers := append(m.removeLocalDefenceModifiers.Get(targetCharacter), handler)
 		m.removeLocalDefenceModifiers.Set(targetCharacter, modifiers)
@@ -191,8 +191,8 @@ func (m *ModifierContext) RemoveLocalDefenceModifier(targetCharacter uint, handl
 	}
 }
 
-func (m *ModifierContext) RemoveLocalChargeModifier(targetCharacter uint, handler modifier.Modifier[ChargeContext]) {
-	var null kv.Map[uint, []modifier.Modifier[ChargeContext]] = nil
+func (m *ModifierContext) RemoveLocalChargeModifier(targetCharacter uint64, handler modifier.Modifier[ChargeContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[ChargeContext]] = nil
 	if m.removeLocalChargeModifiers != null {
 		modifiers := append(m.removeLocalChargeModifiers.Get(targetCharacter), handler)
 		m.removeLocalChargeModifiers.Set(targetCharacter, modifiers)
@@ -203,8 +203,8 @@ func (m *ModifierContext) RemoveLocalChargeModifier(targetCharacter uint, handle
 	}
 }
 
-func (m *ModifierContext) RemoveLocalHealModifier(targetCharacter uint, handler modifier.Modifier[HealContext]) {
-	var null kv.Map[uint, []modifier.Modifier[HealContext]] = nil
+func (m *ModifierContext) RemoveLocalHealModifier(targetCharacter uint64, handler modifier.Modifier[HealContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[HealContext]] = nil
 	if m.removeLocalHealModifiers != null {
 		modifiers := append(m.removeLocalHealModifiers.Get(targetCharacter), handler)
 		m.removeLocalHealModifiers.Set(targetCharacter, modifiers)
@@ -215,8 +215,8 @@ func (m *ModifierContext) RemoveLocalHealModifier(targetCharacter uint, handler 
 	}
 }
 
-func (m *ModifierContext) RemoveLocalCostModifier(targetCharacter uint, handler modifier.Modifier[CostContext]) {
-	var null kv.Map[uint, []modifier.Modifier[CostContext]] = nil
+func (m *ModifierContext) RemoveLocalCostModifier(targetCharacter uint64, handler modifier.Modifier[CostContext]) {
+	var null kv.Map[uint64, []modifier.Modifier[CostContext]] = nil
 	if m.removeLocalCostModifiers != null {
 		modifiers := append(m.removeLocalCostModifiers.Get(targetCharacter), handler)
 		m.removeLocalCostModifiers.Set(targetCharacter, modifiers)
@@ -275,27 +275,27 @@ func (m *ModifierContext) RemoveGlobalCostModifier(handler modifier.Modifier[Cos
 	}
 }
 
-func (m ModifierContext) AddLocalDirectAttackModifiers() kv.Map[uint, []modifier.Modifier[DamageContext]] {
+func (m ModifierContext) AddLocalDirectAttackModifiers() kv.Map[uint64, []modifier.Modifier[DamageContext]] {
 	return m.addLocalDirectAttackModifiers
 }
 
-func (m ModifierContext) AddLocalFinalAttackModifiers() kv.Map[uint, []modifier.Modifier[DamageContext]] {
+func (m ModifierContext) AddLocalFinalAttackModifiers() kv.Map[uint64, []modifier.Modifier[DamageContext]] {
 	return m.addLocalFinalAttackModifiers
 }
 
-func (m ModifierContext) AddLocalDefenceModifiers() kv.Map[uint, []modifier.Modifier[DamageContext]] {
+func (m ModifierContext) AddLocalDefenceModifiers() kv.Map[uint64, []modifier.Modifier[DamageContext]] {
 	return m.addLocalDefenceModifiers
 }
 
-func (m ModifierContext) AddLocalChargeModifiers() kv.Map[uint, []modifier.Modifier[ChargeContext]] {
+func (m ModifierContext) AddLocalChargeModifiers() kv.Map[uint64, []modifier.Modifier[ChargeContext]] {
 	return m.addLocalChargeModifiers
 }
 
-func (m ModifierContext) AddLocalHealModifiers() kv.Map[uint, []modifier.Modifier[HealContext]] {
+func (m ModifierContext) AddLocalHealModifiers() kv.Map[uint64, []modifier.Modifier[HealContext]] {
 	return m.addLocalHealModifiers
 }
 
-func (m ModifierContext) AddLocalCostModifiers() kv.Map[uint, []modifier.Modifier[CostContext]] {
+func (m ModifierContext) AddLocalCostModifiers() kv.Map[uint64, []modifier.Modifier[CostContext]] {
 	return m.addLocalCostModifiers
 }
 
@@ -323,27 +323,27 @@ func (m ModifierContext) AddGlobalCostModifiers() []modifier.Modifier[CostContex
 	return m.addGlobalCostModifiers
 }
 
-func (m ModifierContext) RemoveLocalDirectAttackModifiers() kv.Map[uint, []modifier.Modifier[DamageContext]] {
+func (m ModifierContext) RemoveLocalDirectAttackModifiers() kv.Map[uint64, []modifier.Modifier[DamageContext]] {
 	return m.removeLocalDirectAttackModifiers
 }
 
-func (m ModifierContext) RemoveLocalFinalAttackModifiers() kv.Map[uint, []modifier.Modifier[DamageContext]] {
+func (m ModifierContext) RemoveLocalFinalAttackModifiers() kv.Map[uint64, []modifier.Modifier[DamageContext]] {
 	return m.removeLocalFinalAttackModifiers
 }
 
-func (m ModifierContext) RemoveLocalDefenceModifiers() kv.Map[uint, []modifier.Modifier[DamageContext]] {
+func (m ModifierContext) RemoveLocalDefenceModifiers() kv.Map[uint64, []modifier.Modifier[DamageContext]] {
 	return m.removeLocalDefenceModifiers
 }
 
-func (m ModifierContext) RemoveLocalChargeModifiers() kv.Map[uint, []modifier.Modifier[ChargeContext]] {
+func (m ModifierContext) RemoveLocalChargeModifiers() kv.Map[uint64, []modifier.Modifier[ChargeContext]] {
 	return m.removeLocalChargeModifiers
 }
 
-func (m ModifierContext) RemoveLocalHealModifiers() kv.Map[uint, []modifier.Modifier[HealContext]] {
+func (m ModifierContext) RemoveLocalHealModifiers() kv.Map[uint64, []modifier.Modifier[HealContext]] {
 	return m.removeLocalHealModifiers
 }
 
-func (m ModifierContext) RemoveLocalCostModifiers() kv.Map[uint, []modifier.Modifier[CostContext]] {
+func (m ModifierContext) RemoveLocalCostModifiers() kv.Map[uint64, []modifier.Modifier[CostContext]] {
 	return m.removeLocalCostModifiers
 }
